Refactor update jobs into service module
